Rename OAuthAttributes
